Convert site to use Hugo instead of Lektor
Lektor required python, which requires additional setup on any new system I want
to write on. Hugo, being a go program, is a single executable that requires no
setup beyond downloading the binary, making it easier to get set up on new
systems and making it easier to use in general.